pyproject.toml
[project]
name = "flairsim"
version = "0.1.0"
description = "Drone simulator over FLAIR-HUB aerial imagery for VLM evaluation."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "MS Telecom -- Fil Rouge Team" },
]

dependencies = [
    "numpy>=1.24",
    "rasterio>=1.3",
    "Pillow>=9.0",
]

[project.optional-dependencies]
viewer = [
    "pygame>=2.5",
]
server = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",
]

[project.scripts]
flairsim-server = "flairsim.server.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "httpx>=0.27",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flairsim"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = [
    "D100",  # Missing docstring in public module (we have them)
    "D104",  # Missing docstring in public package
    "D203",  # 1 blank line before class docstring (conflicts with D211)
    "D213",  # Multi-line summary second line (conflicts with D212)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
